requirements.txt
django==4.0.4
psycopg2-binary==2.9.3
djangorestframework==3.12.4
djangorestframework-datatables==0.5.1
django-filter==21.1
requests==2.27.1
pytz==2022.1
python-whois==0.7.3